fix bug kdm002

+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.app/src/main/java/nts/uk/ctx/at/request/app/command/application/remainingnumber/checkfunc/CheckFuncDataCommand.java
@@ -15,5 +15,5 @@ public class CheckFuncDataCommand {
 	private GeneralDate startTime;
 	private GeneralDate endTime;
 	private GeneralDate date;
-	private int maxDay;
+	private double maxDay;
 }
